requirements.txt
aes-everywhere==1.2.10
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.10.0
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
backports.tarfile==1.2.0
bcrypt==4.2.0
billiard==4.2.0
blinker==1.6.2
Brotli==1.0.9
build==1.3.0
CacheControl==0.14.3
celery==5.3.6
celery-types==0.23.0
certifi==2023.5.7
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.1.0
cleo==2.1.0
click==8.1.6
click-didyoumean==0.3.0
click-option-group==0.5.9
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.7
crashtest==0.4.1
cron-descriptor==1.4.3
cryptography==42.0.2
cx_Oracle==8.3.0
Cython==3.1.0
daphne==4.0.0
Deprecated==1.2.18
distlib==0.4.0
Django==5.2.8
django-celery-beat==2.8.1
django-celery-results==2.5.1
django-cors-headers==4.1.0
django-extensions==3.2.3
django-filter==23.2
django-modeltranslation==0.18.10
django-redis==5.4.0
django-storages==1.14.4
django-telegram-logging==1.0.2
django-timezone-field==6.1.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.2.2
dotty-dict==1.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.6
dulwich==0.22.8
ecs-logging==2.2.0
elastic-apm==6.23.0
elastic-transport==8.17.1
elasticsearch==8.17.2
exceptiongroup==1.2.2
Faker==28.4.1
fastjsonschema==2.21.2
filelock==3.19.1
findpython==0.6.3
flake8==7.1.1
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
gevent==23.9.1
geventhttpclient==2.0.9
gitdb==4.0.12
GitPython==3.1.45
graphene==3.3
graphene-django==3.1.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.2.4
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.4
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.1.2
itypes==1.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
Jinja2==3.1.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
keyring==25.6.0
kombu==5.3.5
ldap3==2.9.1
limits==4.2
locust==2.16.1
Markdown==3.4.3
markdown-it-py==4.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
minio==7.2.15
more-itertools==10.7.0
msgpack==1.0.5
numpy==2.0.2
packaging==23.1
pandas==2.3.1
paramiko==3.4.1
pbs-installer==2025.8.14
pdfkit==1.0.0
pkginfo==1.12.1.2
platformdirs==4.3.8
pluggy==1.5.0
poetry-core==2.1.3
prometheus_client==0.21.1
promise==2.3
prompt-toolkit==3.0.43
psutil==5.9.5
psycopg2-binary==2.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.12.1
pycparser==2.21
pycryptodome==3.21.0
pycryptodomex==3.20.0
pydantic==2.12.3
pydantic_core==2.41.4
pyflakes==3.2.0
Pygments==2.19.2
PyJWT==2.7.0
pylogbeat==2.0.1
PyNaCl==1.5.0
pyOpenSSL==24.0.0
pypng==0.20220715.0
pyproject_hooks==1.2.0
pytest==8.3.1
pytest-django==4.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==6.5.0
python-ldap==3.4.4
python-logstash==0.4.8
python-semantic-release==10.4.1
pytz==2023.3
PyYAML==6.0.3
pyzmq==25.1.1
qrcode==7.4.2
RapidFuzz==3.13.0
redis==5.0.1
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
roundrobin==0.0.4
rpds-py==0.27.0
sentry-sdk==1.31.0
service-identity==24.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.13.3
trove-classifiers==2025.8.6.13
Twisted==23.10.0
txaio==23.1.1
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.0.3
vine==5.1.0
virtualenv==20.32.0
wcwidth==0.2.13
Werkzeug==2.3.7
wrapt==1.17.2
xattr==1.2.0
xlsxwriter==3.2.5
zipp==3.23.0
zope.event==5.0
zope.interface==6.0
zstandard==0.23.0
